refactor: 调整UI布局 - 主题按钮和菜单位置优化
